refactor: extract evidence ref resolution and JSONL parsing from parser module
Split parser.py (287 lines, 11 functions) into three focused modules:
- _refs.py (85 lines): CWE URL templates, ref lookups, LLM ref resolution
- _jsonl.py (83 lines): JSONL line parsing, judgment enrichment, file reading
- parser.py (85 lines): evidence assembly, EvidenceContext, re-exports for backward compat
